requirements.txt
flask
flask_restful
python-telegram-bot
requests
beautifulsoup4
lxml